Rename accumulator interface
