+++ CourageScores/ClientApp/src/components/tournaments/competition.ts
@@ -3,7 +3,7 @@ import {GameMatchOptionDto} from "../../interfaces/models/dtos/Game/GameMatchOpt
 import {TournamentSideDto} from "../../interfaces/models/dtos/Game/TournamentSideDto";
 import {any} from "../../helpers/collections";
 import {ILayoutDataForRound} from "./layout";
-import {IMnemonicAccumulator} from "./layout/shared";
+import {IMnemonicGenerator} from "./layout/MnemonicGenerators";
 import {PlayedEngine} from "./layout/PlayedEngine";
 import {UnplayedEngine} from "./layout/UnplayedEngine";
 import {ILayoutEngine} from "./layout/ILayoutEngine";
@@ -11,7 +11,7 @@ import {ILayoutEngine} from "./layout/ILayoutEngine";
 export interface ITournamentLayoutGenerationContext {
     matchOptionDefaults: GameMatchOptionDto;
     getLinkToSide(side: TournamentSideDto): JSX.Element;
-    matchMnemonic?: IMnemonicAccumulator;
+    matchMnemonic?: IMnemonicGenerator;
 }
 
 export function getLayoutData(round: TournamentRoundDto, sides: TournamentSideDto[], context: ITournamentLayoutGenerationContext): ILayoutDataForRound[] {
