+++ CourageScores/ClientApp/src/components/tournaments/layout/MnemonicGenerators.ts
@@ -1,0 +1,12 @@
+export interface IMnemonicGenerator {
+    next(): string;
+}
+
+export function getPrefixIncrementingMnemonicCalculator(prefix: string): IMnemonicGenerator {
+    let index: number = 0;
+    return {
+        next(): string {
+            return prefix + (++index);
+        }
+    };
+}
